requirements.txt
absl-py==2.2.2
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.1
cycler==0.12.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
flatbuffers==25.2.10
fonttools==4.57.0
idna==3.10
jax==0.5.3
jaxlib==0.5.3
kiwisolver==1.4.8
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
mssql-django==1.6
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.6
psutil==7.0.0
pycparser==2.22
pyodbc==5.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyvirtualcam==0.12.1
requests==2.32.4
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
